Fixed error with meshlet buffer size

+++ Engine/QuarantineEngine/Utilities/Geometry/Meshlet.h
@@ -11,12 +11,13 @@ struct MeshletGPU
 {
     glm::vec3 center;
     float radius;
-    int8_t cone_axis[3];
-    int8_t cone_cutoff;
+    float cone_axis[3];
+    float cone_cutoff;
     uint32_t data_offset;
     uint32_t mesh_index;
-    uint8_t vertex_count;
-    uint8_t triangle_count;
+    uint32_t vertex_count;
+    uint32_t triangle_count;
+
 };
 
 class Meshlet
